IGNITE-14435 Refactor PdsConsistentIdProcessor for reusage
